refactor(root): remove compatibility shims after src migration
