--- importer.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/importer.js";
