--- metadata-lint.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/metadata-lint.js";
