+++ package.json
@@ -8,27 +8,27 @@
   "exports": {
     ".": "./index.js",
     "./index.js": "./index.js",
-    "./async-jobs.js": "./async-jobs.js",
-    "./async-progress.js": "./async-progress.js",
-    "./helpers.js": "./helpers.js",
-    "./scene-character-batch.js": "./scene-character-batch.js",
-    "./scrivener-direct.js": "./scrivener-direct.js",
-    "./importer.js": "./importer.js",
-    "./db.js": "./db.js",
-    "./sync.js": "./sync.js",
-    "./git.js": "./git.js",
-    "./world-entity-templates.js": "./world-entity-templates.js",
-    "./workflow-catalogue.js": "./workflow-catalogue.js",
-    "./runtime-diagnostics.js": "./runtime-diagnostics.js",
-    "./metadata-lint.js": "./metadata-lint.js",
-    "./scene-character-normalization.js": "./scene-character-normalization.js",
-    "./review-bundles.js": "./review-bundles.js",
-    "./review-bundles-planner.js": "./review-bundles-planner.js",
-    "./review-bundles-renderer.js": "./review-bundles-renderer.js",
-    "./review-bundles-writer.js": "./review-bundles-writer.js",
-    "./prose-styleguide.js": "./prose-styleguide.js",
-    "./prose-styleguide-drift.js": "./prose-styleguide-drift.js",
-    "./prose-styleguide-skill.js": "./prose-styleguide-skill.js",
+    "./async-jobs.js": "./src/runtime/async-jobs.js",
+    "./async-progress.js": "./src/runtime/async-progress.js",
+    "./helpers.js": "./src/core/helpers.js",
+    "./scene-character-batch.js": "./src/sync/scene-character-batch.js",
+    "./scrivener-direct.js": "./src/sync/scrivener-direct.js",
+    "./importer.js": "./src/sync/importer.js",
+    "./db.js": "./src/core/db.js",
+    "./sync.js": "./src/sync/sync.js",
+    "./git.js": "./src/core/git.js",
+    "./world-entity-templates.js": "./src/world/world-entity-templates.js",
+    "./workflow-catalogue.js": "./src/workflows/workflow-catalogue.js",
+    "./runtime-diagnostics.js": "./src/runtime/runtime-diagnostics.js",
+    "./metadata-lint.js": "./src/sync/metadata-lint.js",
+    "./scene-character-normalization.js": "./src/sync/scene-character-normalization.js",
+    "./review-bundles.js": "./src/review-bundles/review-bundles.js",
+    "./review-bundles-planner.js": "./src/review-bundles/review-bundles-planner.js",
+    "./review-bundles-renderer.js": "./src/review-bundles/review-bundles-renderer.js",
+    "./review-bundles-writer.js": "./src/review-bundles/review-bundles-writer.js",
+    "./prose-styleguide.js": "./src/styleguide/prose-styleguide.js",
+    "./prose-styleguide-drift.js": "./src/styleguide/prose-styleguide-drift.js",
+    "./prose-styleguide-skill.js": "./src/styleguide/prose-styleguide-skill.js",
     "./tools/*": "./src/tools/*",
     "./scripts/*": "./scripts/*",
     "./src/*": "./src/*",
@@ -42,27 +42,6 @@
     "bin/",
     "index.js",
     "src/",
-    "async-jobs.js",
-    "async-progress.js",
-    "helpers.js",
-    "scene-character-batch.js",
-    "scrivener-direct.js",
-    "importer.js",
-    "db.js",
-    "sync.js",
-    "git.js",
-    "world-entity-templates.js",
-    "workflow-catalogue.js",
-    "runtime-diagnostics.js",
-    "metadata-lint.js",
-    "scene-character-normalization.js",
-    "review-bundles.js",
-    "review-bundles-planner.js",
-    "review-bundles-renderer.js",
-    "review-bundles-writer.js",
-    "prose-styleguide.js",
-    "prose-styleguide-drift.js",
-    "prose-styleguide-skill.js",
     "scripts/",
     "README.md",
     "CHANGELOG.md"
