--- prose-styleguide.js
@@ -1,1 +1,0 @@
-export * from "./src/styleguide/prose-styleguide.js";
