--- review-bundles-planner.js
@@ -1,1 +1,0 @@
-export * from "./src/review-bundles/review-bundles-planner.js";
