--- review-bundles-renderer.js
@@ -1,1 +1,0 @@
-export * from "./src/review-bundles/review-bundles-renderer.js";
