--- review-bundles-writer.js
@@ -1,1 +1,0 @@
-export * from "./src/review-bundles/review-bundles-writer.js";
