--- review-bundles.js
@@ -1,1 +1,0 @@
-export * from "./src/review-bundles/review-bundles.js";
