--- runtime-diagnostics.js
@@ -1,1 +1,0 @@
-export * from "./src/runtime/runtime-diagnostics.js";
