--- scene-character-batch.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/scene-character-batch.js";
