--- scene-character-normalization.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/scene-character-normalization.js";
