--- scrivener-direct.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/scrivener-direct.js";
