--- sync.js
@@ -1,1 +1,0 @@
-export * from "./src/sync/sync.js";
