--- workflow-catalogue.js
@@ -1,1 +1,0 @@
-export * from "./src/workflows/workflow-catalogue.js";
