--- world-entity-templates.js
@@ -1,1 +1,0 @@
-export * from "./src/world/world-entity-templates.js";
